Create tweepy_streamer_red.py
This program uses the Twitter API and the Tweepy module to collect tweets. Addressing the big data problem, a combination of methods like filter, split, partition, and other programmatic creativity contribute to this applied human readable solution.

This project would not have been possible if it had not been for Vincent Russo (@CaptainHamptons), who must be credited for his brilliant tutorials.